Refactor: udistrital/sga_documentacion#67 admiciones_mid  Refactorización respuesta e importes
